Implement generate
